changed floats to doubles in the drive controller, fixed a case issue and cleaned up the drive controller a bit

+++ src/controls/wheel_control/include/waypoint_controller/maneuver.h
@@ -3,7 +3,7 @@
 
 struct maneuver
 {
-  float radius, xc, yc, distance;
+  double radius, xc, yc, distance;
 };
 
 #endif
